Function mockito::mock

pub fn mock(method: &str, path: &str) -> Mock

Initializes a mock for the provided method and path.

The mock is registered to the server only after the create() method has been called.

Example

use mockito::mock;

mock("GET", "/");
mock("POST", "/users");
mock("DELETE", "/users?id=1");