Struct mockito::Mock

pub struct Mock {
    // some fields omitted
}

Stores information about a mocked request. Should be initialized via mockito::mock().

Methods

impl Mock

fn match_header(&mut self, field: &str, value: &str) -> &mut Self

Allows matching a particular request header when responding with a mock.

When matching a request, the field letter case is ignored.

Example

use mockito::mock;

mock("GET", "/").match_header("content-type", "application/json");

Like most other Mock methods, it allows chanining:

Example

use mockito::mock;

mock("GET", "/")
  .match_header("content-type", "application/json")
  .match_header("authorization", "password");

fn with_status(&mut self, status: usize) -> &mut Self

Sets the status code of the mock response. The default status code is 200.

Example

use mockito::mock;

mock("GET", "/").with_status(201);

fn with_header(&mut self, field: &str, value: &str) -> &mut Self

Sets a header of the mock response.

Example

use mockito::mock;

mock("GET", "/").with_header("content-type", "application/json");

fn with_body(&mut self, body: &str) -> &mut Self

Sets the body of the mock response. Its Content-Length is handled automatically.

Example

use mockito::mock;

mock("GET", "/").with_body("hello world");

fn with_body_from_file(&mut self, path: &str) -> &mut Self

Sets the body of the mock response from the contents of a file stored under path. Its Content-Length is handled automatically.

Example

use mockito::mock;

mock("GET", "/").with_body_from_file("tests/files/simple.http");

fn create(&self)

Registers the mock to the server - your mock will be served only after calling this method.

Example

use mockito::mock;

mock("GET", "/").with_body("hello world").create();

fn matches(&self, request: &Request) -> bool