Struct mockito::Mock

pub struct Mock { /* fields omitted */ }

Stores information about a mocked request. Should be initialized via mockito::mock().

Methods

impl Mock

fn match_header<M: Into<Matcher>>(self, field: &str, value: M) -> Self

Allows matching a particular request header when responding with a mock.

When matching a request, the field letter case is ignored.

Example

use mockito::mock;

let _m = mock("GET", "/").match_header("content-type", "application/json");

Like most other Mock methods, it allows chanining:

Example

use mockito::mock;

let _m = mock("GET", "/")
  .match_header("content-type", "application/json")
  .match_header("authorization", "password");

fn with_status(self, status: usize) -> Self

Sets the status code of the mock response. The default status code is 200.

Example

use mockito::mock;

let _m = mock("GET", "/").with_status(201);

fn with_header(self, field: &str, value: &str) -> Self

Sets a header of the mock response.

Example

use mockito::mock;

let _m = mock("GET", "/").with_header("content-type", "application/json");

fn with_body(self, body: &str) -> Self

Sets the body of the mock response. Its Content-Length is handled automatically.

Example

use mockito::mock;

let _m = mock("GET", "/").with_body("hello world");

fn with_body_from_file(self, path: &str) -> Self

Sets the body of the mock response from the contents of a file stored under path. Its Content-Length is handled automatically.

Example

use mockito::mock;

let _m = mock("GET", "/").with_body_from_file("tests/files/simple.http");

fn expect(self, hits: usize) -> Self

Sets the expected amount of requests that this mock is supposed to receive. This is only enforced when calling the assert method. Defaults to 1 request.

fn assert(&self)

Asserts that the expected amount of requests (defaults to 1 request) were performed.

fn create(self) -> Self

Registers the mock to the server - your mock will be served only after calling this method.

Example

use mockito::mock;

let _m = mock("GET", "/").with_body("hello world").create();